Explicitly define enum type as string in schema
Many attribute in the request schema are defined with
allowed enum values. But schema does not specify those
enum as string type.

enum in json schema is defined in such a way that it
allows any type not just string so defining each enum type
as string explicitly makes schema more readable and consistent
with other API schema.

Note- this does not change any behavior or does not fix any issue.
This is just for consistency with other schema definition and readability.

Change-Id: I4f6fe5e418c624e0dcbcfb3b6bca64ab3e5c96fe