Make server groups api aware of multiple cells for membership
This teaches the server groups api to query all of the cells that have
an instance claiming to be a member of the group. Instead of polling all
cells, this collects the subset of cells that actually have members,
according to our InstanceMapping data, and then checks for the deleted-ness
in the actual cell database.

This augments the tests to have one instance in each cell, plus an instance
that is not yet scheduled in any cell to give us coverage of all the
possibilities.

This replaces the previously-flawed and now reverted change:

  If571f9e7c9d0ef6265a249ff808d30a24ab462af

Change-Id: Idd2e35bc95ed98ebc0340ff62e109e23c8adcb21