libvirt: configure trust mode for vfs
The PCI framework allows specifying some tags in the whitelist. The
operator who wants to enable 'trust' mode for certains pools of PCI
devices will have to add a "trusted" tag.

[pci]
passthrough_whitelist = {"devname": "eth0",
                         "trusted": "true",
                         "physical_network": "phys0"}

In Neutron the ports will have to be created with specific bindings to
match the tag so Nova will schedule the VM on a host which provides
such PCI devices (assuming that: PciPassthroughFilter is enabled).

neutron port-create <net-id> \
                    --name sriov_port \
                    --vnic-type direct \
                    --binding:profile type=dict trusted=true

Implements blueprint sriov-trusted-vfs
Signed-off-by: Sahid Orentino Ferdjaoui <sahid.ferdjaoui@redhat.com>
Change-Id: I136ab12f2a97ede494af9ca503f160bf1b2aa8fa