Mock timeout in test__get_node_console_with_reset_wait_timeout
This test is failing intermittently because it's waiting for a real
timeout from BackOffLoopingCall which does an exponential backoff
each interval. Sometimes the test times out entirely and fails.

This changes the test to mock BackOffLoopingCall and raise the timeout
exception to simulate a timeout. The only thing not covered by this is
calling of the _wait_state() function, but that is covered by other
unit tests in the file.

Closes-Bug: #1683953

Change-Id: Ie3ff6b0020dca8a5f5292b327aa2491920e338ef