Make target_cell() yield a new context
This makes target_cell() do the thing it was originally intended to do,
which is yield a new context that is targeted at the proper cell. With this,
any objects you create inside the target context manager (assuming you use
the yielded context) will remain targeted after the thing returns.

Related to blueprint cells-aware-api

Change-Id: Iccdf6b80f5fc8adcc8a89ce6ece3f37b6cbcaee2