cinder: add attachment_update method
This adds the attachment_update method to the internal
cinder volume API.

With the new style volume attachments (cinder microversion 3.27),
when attaching a volume to an instance, the API will create
an attachment without a connector which puts the attachment and
volume into 'reserved' status. Once we're on the compute host
and get the connector, we update the attachment which effectively
initializes the connection to the volume backend and marks the
volume as attached to the instance.

Part of blueprint cinder-new-attach-apis

Change-Id: I3f5eb47b15a814291694f20e5b13bcbc16e8a3c0