libvirt: update logic to configure device for scsi controller
When using virtio-scsi it is possible to handle up to 256 disk but
because we do not specifically set the information about the
controller only 6 can be handled.

This commit is fixing the issue by adding the address element to the
disk configuration.

 Conflicts:
	nova/virt/libvirt/driver.py

Closes-Bug: #1686116
Change-Id: I98e53b378cc99747765066001a0b51880543d2dd
(cherry picked from commit fb343c402296c109af49bfca2f909680ca38a72b)