Add `img_hide_hypervisor_id` image property
Image hide_hypervisor_id property helps libvirt set an xml instance file
property, which will hide on guest host KVM hypervisor signature
("KVMKVMKVM\0\0\0").
According to the commit message in QEMU repository [1]:

"The latest Nvidia driver (337.88) specifically checks
for KVM as the hypervisor and reports Code 43 for the
driver  in a Windows guest when found.  Removing or
changing the KVM signature is sufficient for the driver
to load and work."

DocImpact: New feature ``img_hide_hypervisor_id`` image property should be
added in the glance-property-keys page of the cli-reference docs [2].

[1]: http://git.qemu.org/?p=qemu.git;a=commitdiff;h=f522d2a
[2]: https://docs.openstack.org/cli-reference/glance-property-keys.html

Implements: blueprint add-kvm-hidden-feature

Co-Authored-By: Adam Kijak <adam.kijak@corp.ovh.com>

Change-Id: Ie8227fececa40e502aaa39d77de2a1cd0cd72682