Send out notifications when instance tags changed
This patch adds notifications when instance
tags changed, as tag is commonly used for
searching and filtering, and projects like
searchlight depends on the notifications.

To avoid unnecessarily building a notification
payload when notifications are disabled, this
change also adds the if_notifications_enabled
decorator to send_instance_update_notification.

Change-Id: I03c8e8225e51fd80580772752c0b292987e34218
Implements: bp send-tag-notification