Ensure sample policy help text correctly wrapped
On the nova side, we remove any wrapping of the text, but preserve the
paragraph separations. Once the oslo policy generator adds its own line
wrapping, everything looks nice in the sample file.

This change also adds additional line space between the description, the
APIs and the default policy rule, so its a little easier on the eye.

Change-Id: Ib7225309df1a71a2a65625af1fdc9c4007f8a852