Do not log live migration success when it actually failed
During post live migration, if post live migration on destination
fails, then we log a stacktrace but continue to perform cleanup
on the source side. However, at the end of the _post_live_migration
method it was logging that things were successful on the destination
host, which they weren't, which is really confusing when you're trying
to debug the failure and seeing this conflict in the logs.

This patch simply sets a flag if we failed post live migration at
the destination host so we don't log the success message later on
the source host, plus tests to show the flag is set and checked.

Change-Id: I16e70912a13c963031397e66a8553b2c199d50bd
Closes-Bug: #1685340