devref and reno for nova-{api,metadata}-wsgi scripts
This provides a brief explanation of the new nova-api-wsgi [1] and
nova-metadata-wsgi [2] scripts in the Architecture section of the devref
with links to the new doc added to the man pages for the eventlet
scripts.

The nova-api.rst mentioned ec2 so figured best to fix that now
rather than forget about it, despite not being entirely germane.

There is also a reno note that indicates the availability of the new
scripts.

There is a devstack change which is testing the new wsgi scripts as
well as forcing grenade to not use them at
If2d7e363a6541854f2e30c03171bef7a41aff745

[1] I7c4acfaa6c50ac0e4d6de69eb62ec5bbad72ff85
[2] Icb35fe2b94ab02c0ba8ba8129ae18aae0f794756

Change-Id: I351b2af3b256d3031bd2a65feba0495e815f8427
Related-Bug: #1661360