Add compute_nodes_uuid_idx unique index
Since change Ie12acb76ec5affba536c3c45fbb6de35d64aea1b in Ocata
the FilterScheduler is calling the placement service to get a list
of resource providers that can service a request spec, and then
taking that list and querying the compute nodes table with a list
of uuids.

We should have a unique index on the uuid column if we're doing
queries like this, especially somewhere where performance matters
like the scheduler.

As a result of this change, several unit tests that were blindly
creating duplicate compute nodes with the same uuid have to be
fixed to use unique uuids.

Change-Id: I940979d85fc105c9733ae19c406e85debe8fdffb