placement: shared resources when finding providers
This patch implements the query construction for finding resource
providers in a Nova deployment with shared resource providers. A new
_get_all_with_shared() function constructs an advanced SQL expression
that takes into account providers that share resource with other
providers via an aggregate association.

The SQL generated here is pretty complex so I've included ample code
comments showing the logic of how the join expressions are built up,
one on top of the other.

The query plan produced is efficient.  Tested on a database populated
with 10K compute nodes arranged in 100 aggregates each with an
associated storage pool, a MySQL server with minimal RAM (128MB InnoDB
buffer pool size) was able to execute the above in an average of 0.38s,
with the resulting query plan showing indexes being used to satisfy all
join conditions:

http://paste.openstack.org/show/608553/

blueprint: shared-resources-pike

Change-Id: Icd95b2754303a5efd47177eded0baaad56d91677