placement: tests for non-shared with shared
Adds some more functional tests for the _get_all_with_shared() function
that tests an environment with both shared resource providers and
resource providers with "local" resources.

We create a compute node not associated with any aggregate and having
"local" disk storage and verify this compute node is returned along
with compute nodes with shared storage via an aggregate association.

We then allocate against this compute node with "local" disk storage,
consuming all the vCPU and RAM on the compute node (but not disk) and
verify that the node is no longer returned from _get_all_with_shared()

Change-Id: I3e4bc1d0dd5a71668265033dbdf9960f789cded0
blueprint: shared-resources-pike