placement: test for agg association not sharing
Adds yet another functional test scenario that demonstrates an
environment with a set of compute nodes with no local disk connected to
a shared storage provider via an aggregate and a set of compute nodes
with no local disk associated to an aggregate with a provider having
DISK_GB inventory but not decorated with the MISC_SHARES_VIA_AGGREGATE
trait.

This emulates a deployment where you might have a mix of shared and
non-shared storage compute nodes grouped into different host aggregates
and you want fine-grained control over which providers are sharing some
resource with providers in the host aggregate.

Change-Id: Ie71e23b7f51756df4dab2ffb3b7fa938e1a4b475
blueprint: shared-resources-pike