XenAPI: Create linux bridge in dest host during live migration
We found VM can live migrate successfully with XenServer hypervisor,
but after migrating to dest host, we cannot ping the VM's floating
or fixed IP any more. The problem is, during live migration, we only
create the network in dest host which the VM's VIF connect to, but we
don't establish the linux bridge and the corresponding veth pair to
integration bridge br-int.

Closes-Bug: 1677937

Change-Id: I0d69d2ba5b064ef303445892791804608e7f19c5