libvirt: QEMU native LUKS decryption for encrypted volumes
QEMU 2.6 and Libvirt 2.2.0 allow LUKS encrypted RAW files, block devices
and network devices (such as rbd) to be decrypted natively by QEMU. This
change enables the use of this feature within Nova when the appropriate
versions of QEMU and Libvirt are installed and the encryption provider
for the volume is of type 'luks'.

When these conditions are met a Libvirt secret is created when
connecting encrypted volumes to a compute host to hold the LUKS
passphrase used to unlock the volume. The presence of this Libvirt
secret is then used by the volume driver to generate the required
encryption XML for the disk. QEMU is then able to natively read from and
write to the encrypted disk, removing the need for the os-brick supplied
dm-crypt style encryptors, previously used to handle encrypted volumes.

When disconnecting a volume the presence of a Libvirt secret will result
in no attempt being made to detach an os-brick provided encryptor from
the volume. This will only occur when no secret for the volume is found
on the compute host. This will allow encrypted volumes attached prior to
this change to still be detached correctly.

Attempts to swap between volumes while using native QEMU decryption will
be blocked in the same manner as they are when using volumes that do not
provide a local block device. Both use cases still require additional
implementation work in Libvirt before being allowed within Nova.

LibvirtLiveMigrateData and LibvirtLiveMigrateBDMInfo are both extended
to support the following matrix of live migration (LM) scenarios:

- Pike using os-brick encryptors to Queens using os-brick encryptors
- Queens using os-brick encryptors to Queens using os-brick encryptors
- Queens using os-brick encryptors to Queens using native QEMU decrypt
- Queens using native QEMU decrypt to Queens using native QEMU decrypt

A new 'src_supports_native_luks' attribute has been added to
LibvirtLiveMigrateData in Queens to indicate that the source host is
capable of configuring QEMU LUKS decryption for a volume during LM.

The presence of this attribute in migrate_data during pre_live_migration
on the destination is then used to decide how encrypted volumes are
connected on that host ahead of LM starting.

When missing or False the os-brick encryptors will be attached, when
present and True the native QEMU decryption approach will be taken only
if the destination host also supports native LUKS decryption by QEMU.

The UUID of this secret is then stored in the individual
LibvirtLiveMigrateBDMInfo object associated with the volume and passed
back to the source host to be added to the volume configuration prior to
the start of the live migration.

Implements: blueprint libvirt-qemu-native-luks
Change-Id: Ibfa64f18bbd2fb70db7791330ed1a64fe61c1355