Handle network-changed event for a specific port
It turns out that when Neutron sends the "network-changed"
event it has a port in scope and can provide the port ID.

Older versions of Neutron wouldn't send this, but if it's
provided, we can try to optimize the network info cache
refresh and scope it to just that single port, rather than
build the entire cache all over again when the other ports
in the cache may not have changed at all since the last
time the cache was refreshed.

This can be especially beneficial for instances with a
relatively large number of ports which are being migrated.

Depends-On: Ifdaef05208d09ddd9587fed6214cf388e5265ba4

Change-Id: I023b5b1ccb248e68189f62ba0ff75d41093c1f60
Partial-Bug: #1691602
(cherry picked from commit 537f480927293755bd6365068c1abd0c80951bad)