Remove DeviceIsBusy exception
This was added back in Essex with change
Ie18377ba6acd6226612c70fa209185cc579c2d85.

There was never a test added for this libvirtError case
and nothing in upper layers is actually handling DeviceIsBusy,
and since we do the same thing regardless of the type of
exception when a volume attach fails (disconnect the volume),
it's just unnecessary code to have in here, so let's remove it.

Change-Id: I0cdee83c04e17e260a6d5225870be62b6bb4f521