Correct _ensure_console_log_for_instance implementation
_ensure_console_log_for_instance[1] ensures VM console.log existence.

A change[2] updated in order to succeed if the file exists without nova
being able to read it (typically happens when libvirt rewrites uid/gid)
by ignoring EPERM errors.

It seems the method should ignore EACCES errors. Indeed EACCES errors
are raised when an action is not permitted because of insufficient
permissions where EPERM errors when an action is not permitted at all.

[1] nova.virt.libvirt.driver
[2] https://review.openstack.org/392643

Closes-Bug: #1691831
Change-Id: Ifc075a0fd91fc87651fcb306d6439be5369009b6