Skip ceph in grenade live migration job due to restart failures
Chnage I914430d68f64d29932c9409d055b15e4cb384ec4 made the
live migration scripts assume everything is running under systemd,
which is fine for the non-grenade job since devstack on master (pike)
defaults to run everything under systemd.

We missed, however, that the grenade live migration job is starting
from Ocata where screen is used by default, so when we get to the ceph
part of this job in the grenade setup, trying to restart glance-api
via systemctl fails since it's running under screen, not systemd.

For now we'll just skip the ceph live migration setup in the grenade
run until either the bug is fixed or until Queens is our master branch,
at which point the old side for grenade is Pike and running under
systemd.

Change-Id: Ia0ec32dc7cfe744b21b926949c4ab046f9417bc7
Related-Bug: #1691769