Cleanup update_instance cell mapping handling
Targeting the instance mapping cell context is redundant now
that _get_instance stores the targeted context on the
instance._context object when it's retrieved. This happens when
the REST API controller code gets the instance from the compute
API before we ever get to the update_instance method, and
PUT /servers/{server_id} is the only way to get here anyway.

So we can remove all of this older code which was doing an explicit
cell target operation which came before the instance._context
was targeted in _get_instance.

Change-Id: I30b5d4378724b718c6aa8e52ce2f4412d968fc7d