Remove cells v2 transition code from update_instance
Per the TODO in this code, we can remove this transition
code for cells v2 which was added before the instance was
created in conductor. When updating an instance, if the
instance isn't in a cell yet (no id) and the build request was
not found, then either it was already deleted or there is an
instance mapping, because we create the instance mapping before
deleting the build request in conductor. So if the instance
mapping is also not found, then the instance is gone and we can
raise a 404 error.

Change-Id: I6d231320e0825d5c89adfa9caf8b72c59d2a1502