Handle uuid in HostAPI._find_service
This allows looking up a service in the HostAPI by its uuid
value to uniquely identify it in a list of cells. A later
REST API microversion change will rely on this to avoid the
ServiceNotUnique scenario.

Note that the existing service_get_by_id and service_delete
methods are used to minimize the impact to cells v1.

Since we don't have a service mapping table in the API DB we
still need to loop over cells.

A unit test is provided to show that the loop over the cells
is broken once we find the cell that the service lives in
and that the context is targeted to that cell.

Part of blueprint service-hyper-uuid-in-api

Change-Id: Ib82432f125649bc96b386c8f5895813cfbaa57dd