provide interface-scoped nameserver information
DNS nameserver information is subnet-specific, but nova provides this
information globally, which makes it impossible for system
configuration tools to place the nameserver information in the right
place (e.g., in interface configuration files).  This commit modifies
nova to *also* provide the DNS information per-subnet, where it can be
accessed by tools that want to configure the system properly.

Change-Id: I003a25b0d60cb6cd16c3ee1ad1a43910825622be
Closes-bug: #1693265