Have nova.context use super from_context
The nova.context should be a super set of oslo.context in all cases,
especially after we've deserialized on the other side of the RPC
bus. Our current code actually has a completely decoupled from_dict,
which means that as fields are added to oslo.context we don't pick up
any of those, and end up with a potentially broken and fragmented
context on the workers.

This fixes that by using the parent constructor for all the fields we
can, and only explicitly load in a few fields that we also need.

It also simplifies the testing so that we're just testing that our
extra fields end up in the context, and not exact matching everything
in the context, as oslo.context may add important things over time.

Change-Id: Ie683adb36d5e2a736ddbf714524c9c18f3c0d69c