Handle conflict from neutron when addFloatingIP fails
Neutron can raise a Conflict exception when attempting
to associate a floating IP to a server when the fixed
address is already associated to another floating IP.
This has always resulted in a 400 response, however, it
would also trace an ERROR in the nova-api logs, which is
something we shouldn't be doing for an expected type of
failure.

This handles the Conflict in the neutronv2 API client code
and re-raises an exception that the REST API controller code
can handle and return as a 400 without the stacktrace in the
logs.

Change-Id: I27d3241300f75e2aa79a32348a3843e09123cb10
Closes-Bug: #1693576
(cherry picked from commit 452f21183f2f80cc5673ebd3fd3e5daf039caacc)