Send request_id on cinder calls
This changes the constructor so that cinder calls will carry forward
the request_id to the cinder service. ``global_id`` is a magic
property on new oslo.context which is either set to the
global_request_id sent into Nova, or the local request id if it's not
set.

Depends-On: I527026c1c685e5156c410a2b84b2c19cebc71a49

oslo spec I65de8261746b25d45e105394f4eeb95b9cb3bd42

Change-Id: Ic5ee9161cd1174a2dd32b7f155194a7110cc5219