Catch InstanceNotFound when deleting allocations
If the list of allocations that are going to be deleted includes an
instance that can no longer be loaded by get_by_uuid (because
something has already removed it) we need to catch InstanceNotFound
and remove the allocations as planned. Without this change, the
exception is not caught, causing an error and leaving behind the
allocations we don't want to be there.

Change-Id: I204b077c287141a7f5643b2cc0065da2ba395c03
Closes-Bug: #1693899