Changing deleting stale allocations warning to debug
Deleting the allocations is an expected action, and happens
frequently so we don't want it to be a warning. It is useful
as a piece of debug information, however.

Conflicts:
      nova/compute/resource_tracker.py
      nova/tests/unit/compute/test_resource_tracker.py

NOTE(mriedem): The conflict is due to change
I80ba844a6e0fcea89f80aa253d57ac73092773ae in Pike where
this code was all moved to the resource tracker.

Change-Id: Idf88eef036bbe8deca190366f052ab9e355de6e9
Closes-Bug: #1693903
(cherry picked from commit 590ecfb894a7da0c9a16ab9a60638963a10e48db)