Clarify conf/compute.py help text for ListOpts
The compute_monitors setting help gave an example
of providing multiple drivers via the python list
format. As the config was loaded, this caused the
list to be nested within a list.

The non_inheritable_image_properties help also gave
an example in python list format when it should be
a comma-separated list.

Updated the help to show that the compute_monitors
and non_inheritable_image_properties driver settings
should be provided in a comma delimited format if
multiple values are provided.

Change-Id: Ie07892cfde90d1679e941611e1dfbf0c011e1d1b
Closes-Bug: #1693923