[placement] adjust resource provider links by microversion
The resource provider representation includes links to other related
resources. Those links should vary depending on microversion as some
of the related resources did not show up until a later microversion.

This change makes it so 'aggregates' show up at microverison 1.1 and
beyond and traits in microversion 1.6 and beyond. A new gabbit has
been added, resource-provider-links which exercises this. Only the
single resource provider representation is tested because the same
code is used for single and collection representations.

Change-Id: Ife64922ee91a775fabdcf718e6eb0e0c7ddfde7a
Closes-Bug: #1693353