Use microversions for new style volume attachments
This adds the ability to do version discovery with Cinder to
know if we can make requests at a specific microversion, like
3.27 in the case of creating/updating/deleting volume attachments.

We will use this to attempt an attachment_create at 3.27 when
attaching a volume to an instance and if 3.27 is not available
then the calling code will need to handle CinderAPIVersionNotAvailable
and fallback to the old style attachment flow.

For new style attachments, everything will be keyed off the
BlockDeviceMapping.attachment_id field being set if we successfully
created the attachment using 3.27, so the attachment_create method
is really the only one that should care about microversions initially.

Part of blueprint cinder-new-attach-apis

Change-Id: I9b6ab9f71cc58d4514127f5ea61789f187487828