Fix call to driver_detach in remove_volume_connection
DriverVolumeBlockDevice.driver_detach is being called with
the wrong number of arguments - there is no 'connector'
argument in that method. This is a result of refactoring
this code in I7a53e08f3fad6abb27a1d8ad425b4f916341cab3.

The related unit test is pretty brittle so it's easy to
see how this was missed in testing, plus we don't have
integration testing for live migration rollbacks.

This change fixes the call and makes the test a bit less
brittle, but would be cleaner if it could use autospec
although figuring out how to make that work eludes me.

Change-Id: I12088f57f6228460b4810d39e9fca93bd589d70b
Closes-Bug: #1694535