xen: pass Xen console in cmdline
Xen's console is attached to hvc0 in PV guests [1]. If an image does set
it up on its grub configuration, the instance may not boot properly.

Nova is setting the os_cmdline for other hypervisors, so it should do
the same for Xen. This way libvirt will configures the domain properly.

[1] https://wiki.xen.org/wiki/Xen_FAQ_Console

Change-Id: I7600c6e966ab3829185d008077463e9689b9afd5
Closes-Bug: 1691190