libvirt: fix call args to destroy() during live migration rollback
This was missed during I6cca3b5ddd387dac86750be70f49c764a1be2fca
where the migrate_data kwarg was removed from the destroy() method
in the virt drivers.

The existing test was already brittle in that it just asserts that
was passed to the mocked method. This change at least makes the test
a bit less brittle by validating the signature, but there should be
a better way to do this with autospecing and mock, but I couldn't
figure it out.

Change-Id: Iaec2db92b4aa0e6d069c9c5071900e2627683e4e
Closes-Bug: #1694834