Update server create networks API reference description for tags
Artom ruined my day by pointing out that with microversion 2.37
and specifying the enum "auto" for the networks entry in a server
create request, you can't also apply a device tag to the port
that's auto-created for the server.

This change updates the networks API reference parameter description
to note that wrinkle since it's not obvious otherwise without looking
at the API schema.

Change-Id: I1e39aa5a41befb24bb947be060f43f0a96134139