Update live migration to use v3 cinder api
With cinder v3, the new live migration flow is to
create new volume attachments on the destination host
during pre_live_migration, and then after a successful
live migration, delete the attachments on the source host.
If the live migration fails, the attachments on the
destination host will be deleted.

A new dictionary, old_vol_attachment_ids, is added (without
persistence) to migrate_data. This will store the original
attachment_ids of the source host's attachments. If the migrate fails,
the attachment_id in the bdm will be replaced with the old
attachment_id.

An existing unit test, test_pre_live_migration_handles_dict, is
updated to avoid an exception thrown by the new code. A fake
instance is now used instead of a simple dict.

6 new unit tests are added.

Partially Implements: blueprint cinder-new-attach-apis

Depends-On: I9a62b915fc0b9406613c14434793eb52e602df1e

Change-Id: I0bfb11296430dfffe9b091ae7c3a793617bd9d0d