Amend api-ref for multiple networks request
Since there are chances that the guest OS won't honor the order of networks
and assign NIC#1 to the second requested network, we say to our users to
rather use device tagging and metadata querying.
Amending the api-ref to mention explicitly that Nova doesn't support NIC
ordering although there are codepaths for that which try that.

Change-Id: Id134f7745111ccd12695bb041fcce182e833126a
Related-Bug: #1696664