Handle InstanceNotFound when setting password via metadata
When setting an instance password via the metadata service, if the
instance is not found it results in a 500 response to the caller.

This change handles the InstanceNotFound error and returns it as
a 400. Note it's a 400 since the instance uuid is part of the POST
request body, not on the URL path so it's not a 404 response.

Change-Id: I4aa99b563e1a5a87aa3e3dfb28800f107676df92
Partial-Bug: #1696848
(cherry picked from commit c91ee68d493e8c813f4dcc4e46b1794a50f69f32)