Extract custom resource classes from flavors
This patch adds code to look in a flavor's extra_specs for keys
beginning with "resources:", and if found, use them to update the
resources dict sent to placement.

The entry for a custom resource class will look like
"resources:CUSTOM_FOO=1". Additionally, standard classes in a flavor can
be overridden with an entry that looks like: "resources:VCPU=0". If a
standard class is found in extra_specs, it will be used instead of the
amount in the flavor. This is useful for things like Ironic, where an
operator may want to list the amount of RAM, disk, etc. in the flavor,
but use a custom Ironic resource class for doing the actual selection.
An amount for a standard class that is zero will result in that class
being removed from the requested resources dict sent to placement.

DocImpact
We should document the capability and rules in
https://docs.openstack.org/admin-guide/compute-flavors.html#extra-specs.

Blueprint: custom-resource-classes-in-flavors

Change-Id: I84f403fe78e04dd1d099d7d0d1d2925df59e80e7