Centralize compute_node_search_by_hypervisor in os-hypervisors
The search and servers routes have very similar code that is calling
compute_node_search_by_hypervisor and passing the hypervisor
hostname pattern. There is going to be an upcoming change where
we deprecate these APIs and move them into the index() method using
query parameters, so we'll need to use similar code. So centralize
this now so the later change is simpler and can just re-use it.

Part of blueprint service-hyper-uuid-in-api

Change-Id: Ic72b47c11b83ce5151bc34b13eb78af4676dc71a