Stop caching compute nodes in the request
Caching compute nodes in the request was added in change
I73a96db7beb4cc0f017008f81e9f671382ad9105. That was so the
PciHypervisorController could use them, which was an extension
on the hypervisors API.

The os-pci API (and the PciHypervisorController) was removed in
change I9099744264eeec175672d10d04da69648dec1a9d so nothing needs
the compute nodes from the request cache anymore.

Change-Id: I06aacb9d4a8cff8180010c69d3aa32c0492fe2bc