Add datapath type information to OVS vif objects
Added a new parameter to vif objects for modelling the following
OVS port profiles:
  - VIFPortProfileOpenVSwitch
  - VIFPortProfileFPOpenVSwitch

This new parameter named "datapath_type" is a string
coerced variable. The posible values stored in this
field are defined in [1].

This information is required by ``os-vif`` to define or ensure
correctly the datapath type of the OVS bridge when plugging a
port to this bridge. Currently the datapath type is hardcoded
in ``os-vif`` and doesn't work with anything other than the kernel
OVS versions.

[1] https://github.com/openstack/neutron/blob/11.0.0/neutron/plugins/ml2/drivers/openvswitch/agent/common/constants.py#L129

Depends-On: Ia813c39e2917ff373f8e1f85c75fc22d109c94d3
Related-Bug: #1632372

Change-Id: I6194606498bff177d9856ed9f0c86ca0a4eef32f